Update README.md
. Conclusion:

> The Smart Irrigation System reduces manual effort and ensures optimal water usage by automatically controlling irrigation based on soil moisture. This prototype can be further extended using IoT for remote monitoring.




---

. Future Scope :

> Add an IoT dashboard using Blynk / Firebase

Live moisture monitoring on a mobile app

SMS alert when soil is dry





---


Use this:

# Smart Irrigation System using IoT

## ⭐ Project Overview
This project automatically controls the water pump based on soil moisture levels. When the soil becomes dry, the pump turns ON, and when the soil becomes wet, the pump turns OFF. This helps save water and reduces manual monitoring.

Signed-off-by: kiranmai-git11 <kandharikiranmaimiet@gmail.com>